replace styled-components usage with emotion
