+++ frontends/mit-open/package.json
@@ -43,7 +43,6 @@
     "sass": "^1.62.1",
     "sass-loader": "^13.3.0",
     "style-loader": "^3.3.1",
-    "styled-components": "^5.3.5",
     "swc-loader": "^0.2.3",
     "url": "^0.11.0",
     "webpack": "^5.89.0",
