+++ frontends/mit-open/src/pages/Home/HomePageCarousel.tsx
@@ -3,8 +3,7 @@ import styled from "@emotion/styled"
 import { UseQueryResult } from "@tanstack/react-query"
 import ArrowBack from "@mui/icons-material/ArrowBack"
 import ArrowForward from "@mui/icons-material/ArrowForward"
-import { Button, useMuiBreakpointAtLeast } from "ol-design"
-import { TitledCarousel } from "ol-util"
+import { Button, TitledCarousel, useMuiBreakpointAtLeast } from "ol-design"
 import type { PaginatedLearningResourceList } from "api"
 import LearningResourceCard from "../../components/LearningResourceCard"
