+++ frontends/ol-design/src/index.ts
@@ -97,12 +97,14 @@ export { default as FormLabel } from "@mui/material/FormLabel"
 export type { FormLabelProps } from "@mui/material/FormLabel"
 
 export * from "./components/BasicDialog"
+export * from "./components/BannerPage"
 export * from "./components/ButtonLink"
 export * from "./components/ChipLink"
 export * from "./components/FormDialog"
 export * from "./components/RoutedDrawer"
 export * from "./components/SearchInput"
 export * from "./components/SimpleMenu"
+export * from "./components/TitledCarousel"
 export * from "./components/ThemeProvider"
 export * from "./components/LoadingSpinner"
 export * from "./components/RadioChoiceField"
