--- frontends/ol-search-ui/src/styled.d.ts
@@ -1,1 +1,0 @@
-import "ol-util/styled-theme"
