+++ frontends/ol-util/package.json
@@ -11,8 +11,7 @@
   },
   "peerDependencies": {
     "react": "^16.10 || 17 || 18",
-    "react-router": "^5.3.3",
-    "styled-components": "5"
+    "react-router": "^5.3.3"
   },
   "dependencies": {
     "@dnd-kit/core": "^6.0.8",
@@ -21,7 +20,6 @@
     "@faker-js/faker": "^7.3.0",
     "classnames": "^2.3.2",
     "lodash": "^4.17.21",
-    "nuka-carousel": "^5.2.0",
     "qs": "^6.11.0",
     "react-helmet-async": "^1.3.0"
   }
