+++ package.json
@@ -40,7 +40,6 @@
     "@types/jest-when": "^3.5.2",
     "@types/react-dotdotdot": "^1.2.5",
     "@types/react-infinite-scroller": "^1.2.3",
-    "@types/styled-components": "^5.1.25",
     "@types/url-assembler": "^2.1.1",
     "@typescript-eslint/eslint-plugin": "^6.9.1",
     "casual-browserify": "^1.5.19-2",
